condensing files
Progress in condensing files for TDC and ASD GUI functions
