--- GUI/TDC_default_channels.xml
@@ -1,176 +1,0 @@
-<?xml version="1.0" ?>
-<configuration>
-	<mode_set>
-		<enable_new_ttc>0</enable_new_ttc>
-		<enable_master_reset_code>0</enable_master_reset_code>
-		<enable_direct_bunch_reset>0</enable_direct_bunch_reset>
-		<enable_direct_trigger>0</enable_direct_trigger>
-		<auto_roll_over>1</auto_roll_over>
-		<bypass_bcr_distribution>0</bypass_bcr_distribution>
-		<enable_trigger>0</enable_trigger>
-		<channel_data_debug>0</channel_data_debug>
-		<enable_leading>0</enable_leading>
-		<enable_pair>1</enable_pair>
-		<enable_fake_hit>0</enable_fake_hit>
-		<enable_trigger_timeout>0</enable_trigger_timeout>
-		<enable_high_speed>1</enable_high_speed>
-		<enable_legacy>0</enable_legacy>
-		<full_width_res>0</full_width_res>
-		<enable_8b10b>1</enable_8b10b>
-		<enable_insert>0</enable_insert>
-		<enable_error_packet>0</enable_error_packet>
-		<enable_TDC_ID>0</enable_TDC_ID>
-		<enable_error_notify>0</enable_error_notify>
-		<TDC_ID>1111010101010101010</TDC_ID>
-		<width_select>000</width_select>
-		<rising_is_leading>
-			<channel_0>1</channel_0>
-			<channel_1>1</channel_1>
-			<channel_2>1</channel_2>
-			<channel_3>1</channel_3>
-			<channel_4>1</channel_4>
-			<channel_5>1</channel_5>
-			<channel_6>1</channel_6>
-			<channel_7>1</channel_7>
-			<channel_8>1</channel_8>
-			<channel_9>1</channel_9>
-			<channel_10>1</channel_10>
-			<channel_11>1</channel_11>
-			<channel_12>1</channel_12>
-			<channel_13>1</channel_13>
-			<channel_14>1</channel_14>
-			<channel_15>1</channel_15>
-			<channel_16>1</channel_16>
-			<channel_17>1</channel_17>
-			<channel_18>1</channel_18>
-			<channel_19>1</channel_19>
-			<channel_20>1</channel_20>
-			<channel_21>1</channel_21>
-			<channel_22>1</channel_22>
-			<channel_23>1</channel_23>
-		</rising_is_leading>
-		<channel_enable_r>
-			<channel_0>1</channel_0>
-			<channel_1>1</channel_1>
-			<channel_2>1</channel_2>
-			<channel_3>1</channel_3>
-			<channel_4>1</channel_4>
-			<channel_5>1</channel_5>
-			<channel_6>1</channel_6>
-			<channel_7>1</channel_7>
-			<channel_8>1</channel_8>
-			<channel_9>1</channel_9>
-			<channel_10>1</channel_10>
-			<channel_11>1</channel_11>
-			<channel_12>1</channel_12>
-			<channel_13>1</channel_13>
-			<channel_14>1</channel_14>
-			<channel_15>1</channel_15>
-			<channel_16>1</channel_16>
-			<channel_17>1</channel_17>
-			<channel_18>1</channel_18>
-			<channel_19>1</channel_19>
-			<channel_20>1</channel_20>
-			<channel_21>1</channel_21>
-			<channel_22>1</channel_22>
-			<channel_23>1</channel_23>
-		</channel_enable_r>
-		<channel_enable_f>
-			<channel_0>1</channel_0>
-			<channel_1>1</channel_1>
-			<channel_2>1</channel_2>
-			<channel_3>1</channel_3>
-			<channel_4>1</channel_4>
-			<channel_5>1</channel_5>
-			<channel_6>1</channel_6>
-			<channel_7>1</channel_7>
-			<channel_8>1</channel_8>
-			<channel_9>1</channel_9>
-			<channel_10>1</channel_10>
-			<channel_11>1</channel_11>
-			<channel_12>1</channel_12>
-			<channel_13>1</channel_13>
-			<channel_14>1</channel_14>
-			<channel_15>1</channel_15>
-			<channel_16>1</channel_16>
-			<channel_17>1</channel_17>
-			<channel_18>1</channel_18>
-			<channel_19>1</channel_19>
-			<channel_20>1</channel_20>
-			<channel_21>1</channel_21>
-			<channel_22>1</channel_22>
-			<channel_23>1</channel_23>
-		</channel_enable_f>
-	</mode_set>
-	<internal_counter>
-		<combine_time_out_config>0000101000</combine_time_out_config>
-		<fake_hit_time_interval>000100000000</fake_hit_time_interval>
-		<syn_packet_number>111111111111</syn_packet_number>
-		<roll_over>111111111111</roll_over>
-		<coarse_count_offset>000000000000</coarse_count_offset>
-		<bunch_offset>111110011100</bunch_offset>
-		<event_offset>000000000000</event_offset>
-		<match_window>000000011111</match_window>
-	</internal_counter>
-	<fine_time_lut>
-		<fine_sel>0011</fine_sel>
-		<lut0>00</lut0>
-		<lut1>01</lut1>
-		<lut2>10</lut2>
-		<lut3>01</lut3>
-		<lut4>11</lut4>
-		<lut5>00</lut5>
-		<lut6>10</lut6>
-		<lut7>10</lut7>
-		<lut8>00</lut8>
-		<lut9>00</lut9>
-		<luta>00</luta>
-		<lutb>01</lutb>
-		<lutc>11</lutc>
-		<lutd>00</lutd>
-		<lute>11</lute>
-		<lutf>00</lutf>
-	</fine_time_lut>
-	<reset_option>
-		<rst_ePLL>0</rst_ePLL>
-		<reset_jtag_in>0</reset_jtag_in>
-		<event_reset_jtag_in>0</event_reset_jtag_in>
-		<chnl_fifo_overflow_clear>0</chnl_fifo_overflow_clear>
-		<debug_port_select>0000</debug_port_select>
-	</reset_option>
-	<ePll_option>
-		<phase_clk160>01000</phase_clk160>
-		<phase_clk320_0>0100</phase_clk320_0>
-		<phase_clk320_1>0000</phase_clk320_1>
-		<phase_clk320_2>0010</phase_clk320_2>
-		<ePllRes>0010</ePllRes>
-		<ePllIcp>0100</ePllIcp>
-		<ePllCap>10</ePllCap>
-	</ePll_option>
-	<TDC_status/>
-	<data_rate>
-		<run_320_rate>0</run_320_rate>
-		<run_160_rate>1</run_160_rate>
-		<run_80_rate>0</run_80_rate>
-	</data_rate>
-	<trigger_mode>
-		<enable_trigger>1</enable_trigger>
-	</trigger_mode>
-	<formatting>
-		<pair_mode>0</pair_mode>
-		<edge_mode>1</edge_mode>
-		<double_edge_mode>0</double_edge_mode>
-		<pair_full_width_mode>0</pair_full_width_mode>
-		<debug_mode>0</debug_mode>
-		<TDC_ID_mode>0</TDC_ID_mode>
-	</formatting>
-	<edge_type>
-		<leading_type>1</leading_type>
-	</edge_type>
-	<phase_clk160>
-		<phase_clk160_value>7</phase_clk160_value>
-	</phase_clk160>
-	<phase_clk320_1>
-		<phase_clk320_1_value>5</phase_clk320_1_value>
-	</phase_clk320_1>
-</configuration>
