+++ GUI/TDC_what.xml
@@ -1,3 +1,4 @@
+<?xml version="1.0" ?>
 <configuration>
 	<mode_set>
 		<enable_new_ttc>0</enable_new_ttc>
@@ -11,17 +12,6 @@
 		<enable_leading>0</enable_leading>
 		<enable_pair>1</enable_pair>
 		<enable_fake_hit>0</enable_fake_hit>
-		<enable_trigger_timeout>0</enable_trigger_timeout>
-		<enable_high_speed>1</enable_high_speed>
-		<enable_legacy>0</enable_legacy>
-		<full_width_res>0</full_width_res>
-		<enable_8b10b>1</enable_8b10b>
-		<enable_insert>0</enable_insert>
-		<enable_error_packet>0</enable_error_packet>
-		<enable_TDC_ID>0</enable_TDC_ID>
-		<enable_error_notify>0</enable_error_notify>
-		<TDC_ID>1111010101010101010</TDC_ID>
-		<width_select>000</width_select>
 		<rising_is_leading>
 			<channel_0>1</channel_0>
 			<channel_1>1</channel_1>
@@ -100,6 +90,17 @@
 			<channel_22>1</channel_22>
 			<channel_23>1</channel_23>
 		</channel_enable_f>
+		<TDC_ID>1111010101010101010</TDC_ID>
+		<enable_trigger_timeout>0</enable_trigger_timeout>
+		<enable_high_speed>1</enable_high_speed>
+		<enable_legacy>0</enable_legacy>
+		<full_width_res>0</full_width_res>
+		<width_select>000</width_select>
+		<enable_8b10b>1</enable_8b10b>
+		<enable_insert>0</enable_insert>
+		<enable_error_packet>0</enable_error_packet>
+		<enable_TDC_ID>0</enable_TDC_ID>
+		<enable_error_notify>0</enable_error_notify>
 	</mode_set>
 	<internal_counter>
 		<combine_time_out_config>0000101000</combine_time_out_config>
@@ -146,30 +147,30 @@
 		<ePllIcp>0100</ePllIcp>
 		<ePllCap>10</ePllCap>
 	</ePll_option>
-	<TDC_status />
+	<TDC_status/>
 	<data_rate>
-		<run_320_rate>1</run_320_rate>
-		<run_160_rate>0</run_160_rate>
+		<run_320_rate>0</run_320_rate>
+		<run_160_rate>1</run_160_rate>
 		<run_80_rate>0</run_80_rate>
 	</data_rate>
 	<trigger_mode>
-		<enable_trigger>0</enable_trigger>
+		<enable_trigger>1</enable_trigger>
 	</trigger_mode>
 	<formatting>
-		<pair_mode>1</pair_mode>
-		<edge_mode>0</edge_mode>
+		<pair_mode>0</pair_mode>
+		<edge_mode>1</edge_mode>
 		<double_edge_mode>0</double_edge_mode>
 		<pair_full_width_mode>0</pair_full_width_mode>
 		<debug_mode>0</debug_mode>
 		<TDC_ID_mode>0</TDC_ID_mode>
 	</formatting>
 	<edge_type>
-		<leading_type>0</leading_type>
+		<leading_type>1</leading_type>
 	</edge_type>
 	<phase_clk160>
-		<phase_clk160_value>8</phase_clk160_value>
+		<phase_clk160_value>7</phase_clk160_value>
 	</phase_clk160>
 	<phase_clk320_1>
-		<phase_clk320_1_value>0</phase_clk320_1_value>
+		<phase_clk320_1_value>5</phase_clk320_1_value>
 	</phase_clk320_1>
 </configuration>
