+++ GUI/control_0_popup_condensed.py
@@ -7,9 +7,7 @@
 # WARNING! All changes made in this file will be lost!
 
 ######## control0 popup script #########
-
 from PyQt5 import QtCore, QtGui, QtWidgets
-
 
 class Ui_Dialog(object):
 
@@ -23,35 +21,29 @@ class Ui_Dialog(object):
         #Main layout
         self.verticalLayoutWidget = QtWidgets.QWidget(Dialog)
         self.verticalLayoutWidget.setGeometry(QtCore.QRect(10, 20, 311, 151))
-        self.verticalLayoutWidget.setObjectName("verticalLayoutWidget")
         self.verticalLayout = QtWidgets.QVBoxLayout(self.verticalLayoutWidget)
         self.verticalLayout.setContentsMargins(0, 0, 0, 0)
-        self.verticalLayout.setObjectName("verticalLayout")
 
         #rst_ePLL checkbox
         self.checkBox = QtWidgets.QCheckBox(self.verticalLayoutWidget)
-        self.checkBox.setObjectName("checkBox")
         self.verticalLayout.addWidget(self.checkBox)
         if self.TDC_inst.rst_ePLL[0] == '1':
             self.checkBox.setChecked(True)
 
         #reset_jtag_in checkbox
         self.checkBox_2 = QtWidgets.QCheckBox(self.verticalLayoutWidget)
-        self.checkBox_2.setObjectName("checkBox_2")
         self.verticalLayout.addWidget(self.checkBox_2)
         if self.TDC_inst.reset_jtag_in[0] == '1':
             self.checkBox_2.setChecked(True)
 
         #event_reset_jtag_in checkbox
         self.checkBox_3 = QtWidgets.QCheckBox(self.verticalLayoutWidget)
-        self.checkBox_3.setObjectName("checkBox_3")
         self.verticalLayout.addWidget(self.checkBox_3)
         if self.TDC_inst.event_reset_jtag_in[0] == '1':
             self.checkBox_3.setChecked(True)
 
         #chnl_fifo_overflow_clear checkbox
         self.checkBox_4 = QtWidgets.QCheckBox(self.verticalLayoutWidget)
-        self.checkBox_4.setObjectName("checkBox_4")
         self.verticalLayout.addWidget(self.checkBox_4)
         if self.TDC_inst.chnl_fifo_overflow_clear[0] == '1':
             self.checkBox_4.setChecked(True)
@@ -59,14 +51,11 @@ class Ui_Dialog(object):
         #layout for debug_port_select
         self.horizontalLayoutWidget = QtWidgets.QWidget(Dialog)
         self.horizontalLayoutWidget.setGeometry(QtCore.QRect(10, 180, 311, 61))
-        self.horizontalLayoutWidget.setObjectName("horizontalLayoutWidget")
         self.horizontalLayout = QtWidgets.QHBoxLayout(self.horizontalLayoutWidget)
         self.horizontalLayout.setContentsMargins(0, 0, 0, 0)
-        self.horizontalLayout.setObjectName("horizontalLayout")
 
         #debug_port_select
         self.lineEdit = QtWidgets.QLineEdit(self.horizontalLayoutWidget)
-        self.lineEdit.setObjectName("lineEdit")
         self.horizontalLayout.addWidget(self.lineEdit)
         self.lineEdit.setText(self.TDC_inst.debug_port_select[0])
         self.label = QtWidgets.QLabel(self.horizontalLayoutWidget)
@@ -76,32 +65,26 @@ class Ui_Dialog(object):
         # bottom button stuff
         self.horizontalLayoutWidget_6 = QtWidgets.QWidget(Dialog)
         self.horizontalLayoutWidget_6.setGeometry(QtCore.QRect(10, 250, 311, 41))
-        self.horizontalLayoutWidget_6.setObjectName("horizontalLayoutWidget_6")
         self.horizontalLayout_6 = QtWidgets.QHBoxLayout(self.horizontalLayoutWidget_6)
         self.horizontalLayout_6.setContentsMargins(0, 0, 0, 0)
-        self.horizontalLayout_6.setObjectName("horizontalLayout_6")
 
         #Apply button
         self.pushButton = QtWidgets.QPushButton(self.horizontalLayoutWidget_6)
         self.pushButton.setObjectName("pushButton")
         self.horizontalLayout_6.addWidget(self.pushButton)
         self.pushButton.clicked.connect(self.apply_button)
-        #self.pushButton.clicked.connect(self.apply_button_mes)
-        #self.pushButton.clicked.connect(Dialog.reject)
 
         #OK button
         self.pushButton_2 = QtWidgets.QPushButton(self.horizontalLayoutWidget_6)
         self.pushButton_2.setObjectName("pushButton_2")
         self.horizontalLayout_6.addWidget(self.pushButton_2)
-        self.pushButton_2.clicked.connect(self.apply_button)
-        #self.pushButton_2.clicked.connect(self.OK_button_mes)
+        self.pushButton_2.clicked.connect(self.OK_button)
         self.pushButton_2.clicked.connect(Dialog.reject)
 
         #Cancel button
         self.pushButton_3 = QtWidgets.QPushButton(self.horizontalLayoutWidget_6)
         self.pushButton_3.setObjectName("pushButton_3")
         self.horizontalLayout_6.addWidget(self.pushButton_3)
-        #self.pushButton_3.clicked.connect(self.cancel_button_mes)
         self.pushButton_3.clicked.connect(Dialog.reject)
 
         self.retranslateUi(Dialog)
@@ -119,9 +102,16 @@ class Ui_Dialog(object):
         self.pushButton_2.setText(_translate("Dialog", "OK"))
         self.pushButton_3.setText(_translate("Dialog", "Cancel"))
 
-
-    #Function for the apply/OK buttons
+    #apply button
     def apply_button(self):
+        self.save()
+
+    #OK button
+    def OK_button(self):
+        self.save()
+
+    #Function to save variables
+    def save(self):
         #rst_ePLL
         if self.checkBox.isChecked() == True:
             self.TDC_inst.rst_ePLL[0] = '1'
@@ -152,11 +142,11 @@ class Ui_Dialog(object):
         #call update_control_0 function
         self.TDC_inst.update_control_0()
 
-if __name__ == "__main__":
-    import sys
-    app = QtWidgets.QApplication(sys.argv)
-    Dialog = QtWidgets.QDialog()
-    ui = Ui_Dialog()
-    ui.setupUi(Dialog)
-    Dialog.show()
-    sys.exit(app.exec_())
+# if __name__ == "__main__":
+#     import sys
+#     app = QtWidgets.QApplication(sys.argv)
+#     Dialog = QtWidgets.QDialog()
+#     ui = Ui_Dialog()
+#     ui.setupUi(Dialog)
+#     Dialog.show()
+#     sys.exit(app.exec_())
