+++ GUI/setup_2_popup_condensed.py
@@ -8,7 +8,6 @@
 
 
 ################ setup2 popup script ###################
-
 from PyQt5 import QtCore, QtGui, QtWidgets
 
 
@@ -21,6 +20,7 @@ class Ui_Dialog(object):
         Dialog.setObjectName("Dialog")
         Dialog.resize(347, 500)
 
+
         #Main grid layout
         self.gridLayoutWidget = QtWidgets.QWidget(Dialog)
         self.gridLayoutWidget.setGeometry(QtCore.QRect(19, 19, 301, 401))
@@ -31,9 +31,7 @@ class Ui_Dialog(object):
 
         #fine_sel
         self.lineEdit = QtWidgets.QLineEdit(self.gridLayoutWidget)
-        self.lineEdit.setObjectName("lineEdit")
         self.gridLayout.addWidget(self.lineEdit, 0, 1, 1, 1)
-        #self.lineEdit.textChanged.connect(self.changed_line1)
         self.lineEdit.setText(self.TDC_inst.fine_sel[0])
         self.label = QtWidgets.QLabel(self.gridLayoutWidget)
         self.label.setObjectName("label")
@@ -41,9 +39,7 @@ class Ui_Dialog(object):
 
         #lut0
         self.lineEdit_2 = QtWidgets.QLineEdit(self.gridLayoutWidget)
-        #self.lineEdit_2.setObjectName("lineEdit_2")
         self.gridLayout.addWidget(self.lineEdit_2, 1, 1, 1, 1)
-        #self.lineEdit_2.textChanged.connect(self.changed_line2)
         self.lineEdit_2.setText(self.TDC_inst.lut0[0])
         self.label_2 = QtWidgets.QLabel(self.gridLayoutWidget)
         self.label_2.setObjectName("label_2")
@@ -51,9 +47,7 @@ class Ui_Dialog(object):
 
         # lut1
         self.lineEdit_3 = QtWidgets.QLineEdit(self.gridLayoutWidget)
-        #self.lineEdit_3.setObjectName("lineEdit_3")
         self.gridLayout.addWidget(self.lineEdit_3, 2, 1, 1, 1)
-        #self.lineEdit_3.textChanged.connect(self.changed_line3)
         self.lineEdit_3.setText(self.TDC_inst.lut1[0])
         self.label_3 = QtWidgets.QLabel(self.gridLayoutWidget)
         self.label_3.setObjectName("label_3")
@@ -61,9 +55,7 @@ class Ui_Dialog(object):
 
         # lut2
         self.lineEdit_4 = QtWidgets.QLineEdit(self.gridLayoutWidget)
-        self.lineEdit_4.setObjectName("lineEdit_4")
         self.gridLayout.addWidget(self.lineEdit_4, 3, 1, 1, 1)
-        #self.lineEdit_4.textChanged.connect(self.changed_line4)
         self.lineEdit_4.setText(self.TDC_inst.lut2[0])
         self.label_4 = QtWidgets.QLabel(self.gridLayoutWidget)
         self.label_4.setObjectName("label_4")
@@ -71,9 +63,7 @@ class Ui_Dialog(object):
 
         # lut3
         self.lineEdit_5 = QtWidgets.QLineEdit(self.gridLayoutWidget)
-        self.lineEdit_5.setObjectName("lineEdit_5")
         self.gridLayout.addWidget(self.lineEdit_5, 4, 1, 1, 1)
-        # self.lineEdit_4.textChanged.connect(self.changed_line4)
         self.lineEdit_5.setText(self.TDC_inst.lut3[0])
         self.label_5 = QtWidgets.QLabel(self.gridLayoutWidget)
         self.label_5.setObjectName("label_5")
@@ -81,9 +71,7 @@ class Ui_Dialog(object):
 
         # lut4
         self.lineEdit_6 = QtWidgets.QLineEdit(self.gridLayoutWidget)
-        #self.lineEdit_6.setObjectName("lineEdit_6")
         self.gridLayout.addWidget(self.lineEdit_6, 5, 1, 1, 1)
-        # self.lineEdit_4.textChanged.connect(self.changed_line4)
         self.lineEdit_6.setText(self.TDC_inst.lut4[0])
         self.label_6 = QtWidgets.QLabel(self.gridLayoutWidget)
         self.label_6.setObjectName("label_6")
@@ -92,7 +80,6 @@ class Ui_Dialog(object):
         # lut5
         self.lineEdit_7 = QtWidgets.QLineEdit(self.gridLayoutWidget)
         self.gridLayout.addWidget(self.lineEdit_7, 6, 1, 1, 1)
-        # self.lineEdit_4.textChanged.connect(self.changed_line4)
         self.lineEdit_7.setText(self.TDC_inst.lut5[0])
         self.label_7 = QtWidgets.QLabel(self.gridLayoutWidget)
         self.label_7.setObjectName("label_7")
@@ -101,7 +88,6 @@ class Ui_Dialog(object):
         # lut6
         self.lineEdit_8 = QtWidgets.QLineEdit(self.gridLayoutWidget)
         self.gridLayout.addWidget(self.lineEdit_8, 7, 1, 1, 1)
-        # self.lineEdit_4.textChanged.connect(self.changed_line4)
         self.lineEdit_8.setText(self.TDC_inst.lut6[0])
         self.label_8 = QtWidgets.QLabel(self.gridLayoutWidget)
         self.label_8.setObjectName("label_8")
@@ -110,7 +96,6 @@ class Ui_Dialog(object):
         # lut7
         self.lineEdit_9 = QtWidgets.QLineEdit(self.gridLayoutWidget)
         self.gridLayout.addWidget(self.lineEdit_9, 0, 3, 1, 1)
-        # self.lineEdit_4.textChanged.connect(self.changed_line4)
         self.lineEdit_9.setText(self.TDC_inst.lut7[0])
         self.label_9 = QtWidgets.QLabel(self.gridLayoutWidget)
         self.label_9.setObjectName("label_9")
@@ -119,7 +104,6 @@ class Ui_Dialog(object):
         # lut8
         self.lineEdit_10 = QtWidgets.QLineEdit(self.gridLayoutWidget)
         self.gridLayout.addWidget(self.lineEdit_10, 1, 3, 1, 1)
-        # self.lineEdit_4.textChanged.connect(self.changed_line4)
         self.lineEdit_10.setText(self.TDC_inst.lut8[0])
         self.label_10 = QtWidgets.QLabel(self.gridLayoutWidget)
         self.label_10.setObjectName("label_10")
@@ -128,7 +112,6 @@ class Ui_Dialog(object):
         # lut9
         self.lineEdit_11 = QtWidgets.QLineEdit(self.gridLayoutWidget)
         self.gridLayout.addWidget(self.lineEdit_11, 2, 3, 1, 1)
-        # self.lineEdit_4.textChanged.connect(self.changed_line4)
         self.lineEdit_11.setText(self.TDC_inst.lut9[0])
         self.label_11 = QtWidgets.QLabel(self.gridLayoutWidget)
         self.label_11.setObjectName("label_11")
@@ -137,8 +120,8 @@ class Ui_Dialog(object):
         # luta
         self.lineEdit_12 = QtWidgets.QLineEdit(self.gridLayoutWidget)
         self.gridLayout.addWidget(self.lineEdit_12, 3, 3, 1, 1)
-        # self.lineEdit_4.textChanged.connect(self.changed_line4)
-        self.lineEdit_12.setText(self.TDC_inst.luta[0])
+        # self.lineEdit_12.setText(self.TDC_inst.luta[0])
+        self.lineEdit_12.setText(self.TDC_inst.chnl_decode[11][0])
         self.label_12 = QtWidgets.QLabel(self.gridLayoutWidget)
         self.label_12.setObjectName("label_12")
         self.gridLayout.addWidget(self.label_12, 3, 2, 1, 1)
@@ -146,8 +129,8 @@ class Ui_Dialog(object):
         # lutb
         self.lineEdit_13 = QtWidgets.QLineEdit(self.gridLayoutWidget)
         self.gridLayout.addWidget(self.lineEdit_13, 4, 3, 1, 1)
-        # self.lineEdit_4.textChanged.connect(self.changed_line4)
-        self.lineEdit_13.setText(self.TDC_inst.lutb[0])
+        # self.lineEdit_13.setText(self.TDC_inst.lutb[0])
+        self.lineEdit_13.setText(self.TDC_inst.chnl_decode[12][0])
         self.label_13 = QtWidgets.QLabel(self.gridLayoutWidget)
         self.label_13.setObjectName("label_13")
         self.gridLayout.addWidget(self.label_13, 4, 2, 1, 1)
@@ -155,8 +138,8 @@ class Ui_Dialog(object):
         # lutc
         self.lineEdit_14 = QtWidgets.QLineEdit(self.gridLayoutWidget)
         self.gridLayout.addWidget(self.lineEdit_14, 5, 3, 1, 1)
-        # self.lineEdit_4.textChanged.connect(self.changed_line4)
-        self.lineEdit_14.setText(self.TDC_inst.lutc[0])
+        # self.lineEdit_14.setText(self.TDC_inst.lutc[0])
+        self.lineEdit_14.setText(self.TDC_inst.chnl_decode[13][0])
         self.label_14 = QtWidgets.QLabel(self.gridLayoutWidget)
         self.label_14.setObjectName("label_14")
         self.gridLayout.addWidget(self.label_14, 5, 2, 1, 1)
@@ -164,7 +147,6 @@ class Ui_Dialog(object):
         # lutd
         self.lineEdit_15 = QtWidgets.QLineEdit(self.gridLayoutWidget)
         self.gridLayout.addWidget(self.lineEdit_15, 6, 3, 1, 1)
-        # self.lineEdit_4.textChanged.connect(self.changed_line4)
         self.lineEdit_15.setText(self.TDC_inst.lutd[0])
         self.label_15 = QtWidgets.QLabel(self.gridLayoutWidget)
         self.label_15.setObjectName("label_15")
@@ -173,7 +155,6 @@ class Ui_Dialog(object):
         # lute
         self.lineEdit_16 = QtWidgets.QLineEdit(self.gridLayoutWidget)
         self.gridLayout.addWidget(self.lineEdit_16, 7, 3, 1, 1)
-        # self.lineEdit_4.textChanged.connect(self.changed_line4)
         self.lineEdit_16.setText(self.TDC_inst.lute[0])
         self.label_16 = QtWidgets.QLabel(self.gridLayoutWidget)
         self.label_16.setObjectName("label_16")
@@ -182,7 +163,6 @@ class Ui_Dialog(object):
         # lutf
         self.lineEdit_17 = QtWidgets.QLineEdit(self.gridLayoutWidget)
         self.gridLayout.addWidget(self.lineEdit_17, 8, 3, 1, 1)
-        # self.lineEdit_4.textChanged.connect(self.changed_line4)
         self.lineEdit_17.setText(self.TDC_inst.lutf[0])
         self.label_17 = QtWidgets.QLabel(self.gridLayoutWidget)
         self.label_17.setObjectName("label_17")
@@ -191,32 +171,26 @@ class Ui_Dialog(object):
         # bottom button stuff
         self.horizontalLayoutWidget_6 = QtWidgets.QWidget(Dialog)
         self.horizontalLayoutWidget_6.setGeometry(QtCore.QRect(20, 440, 301, 31))
-        self.horizontalLayoutWidget_6.setObjectName("horizontalLayoutWidget_6")
         self.horizontalLayout_6 = QtWidgets.QHBoxLayout(self.horizontalLayoutWidget_6)
         self.horizontalLayout_6.setContentsMargins(0, 0, 0, 0)
-        self.horizontalLayout_6.setObjectName("horizontalLayout_6")
 
         #Apply button
         self.pushButton = QtWidgets.QPushButton(self.horizontalLayoutWidget_6)
         self.pushButton.setObjectName("pushButton")
         self.horizontalLayout_6.addWidget(self.pushButton)
         self.pushButton.clicked.connect(self.apply_button)
-        #self.pushButton.clicked.connect(self.apply_button_mes)
-        #self.pushButton.clicked.connect(Dialog.reject)
 
         #OK button
         self.pushButton_2 = QtWidgets.QPushButton(self.horizontalLayoutWidget_6)
         self.pushButton_2.setObjectName("pushButton_2")
         self.horizontalLayout_6.addWidget(self.pushButton_2)
-        self.pushButton_2.clicked.connect(self.apply_button)
-        #self.pushButton_2.clicked.connect(self.OK_button_mes)
+        self.pushButton_2.clicked.connect(self.OK_button)
         self.pushButton_2.clicked.connect(Dialog.reject)
 
         #Cancel button
         self.pushButton_3 = QtWidgets.QPushButton(self.horizontalLayoutWidget_6)
         self.pushButton_3.setObjectName("pushButton_3")
         self.horizontalLayout_6.addWidget(self.pushButton_3)
-        #self.pushButton_3.clicked.connect(self.cancel_button_mes)
         self.pushButton_3.clicked.connect(Dialog.reject)
 
         self.retranslateUi(Dialog)
@@ -247,7 +221,17 @@ class Ui_Dialog(object):
         self.pushButton_2.setText(_translate("Dialog", "OK"))
         self.pushButton_3.setText(_translate("Dialog", "Cancel"))
 
+    #apply button
     def apply_button(self):
+        self.save()
+
+    #OK button
+    def OK_button(self):
+        self.save()
+        # print(self.TDC_inst.chnl_decode[0][0])
+
+    #save changes
+    def save(self):
         self.TDC_inst.fine_sel[0] = self.lineEdit.text()
         self.TDC_inst.lut0[0] = self.lineEdit_2.text()
         self.TDC_inst.lut1[0] = self.lineEdit_3.text()
@@ -269,11 +253,11 @@ class Ui_Dialog(object):
         #Call update_setup_2 function
         self.TDC_inst.update_setup_2()
 
-if __name__ == "__main__":
-    import sys
-    app = QtWidgets.QApplication(sys.argv)
-    Dialog = QtWidgets.QDialog()
-    ui = Ui_Dialog()
-    ui.setupUi(Dialog)
-    Dialog.show()
-    sys.exit(app.exec_())
+# if __name__ == "__main__":
+#     import sys
+#     app = QtWidgets.QApplication(sys.argv)
+#     Dialog = QtWidgets.QDialog()
+#     ui = Ui_Dialog()
+#     ui.setupUi(Dialog)
+#     Dialog.show()
+#     sys.exit(app.exec_())
